#29812 RSTAB ecc sync fix
